+++ classes.py
@@ -1,0 +1,7 @@
+class Process:
+        
+    def __init__(self, path, process):
+        self.path = path
+        self.process = process
+        self.timesRestarted = 0
+
